refactor backend controllers
